+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - NinjaTerm version number displayed in top-right of screen.
 - Added shorthand serial port config. display in the bottom toolbar of app.
 - Added logo to the app toolbar.
+- Added a "Ko-Fi" donate button to the app toolbar.
 
 ### Changed